PF Improve and modularize carousel controllers
